+++ src/main/php/peer/server/PreforkingServer.class.php
@@ -41,20 +41,6 @@ class PreforkingServer extends Server implements Traceable {
   }
 
   /**
-   * Signal handler
-   *
-   * @param   int sig
-   */
-  public function handleSignal($sig) {
-    $this->cat && $this->cat->debugf('Received signal %d in pid %d', $sig, getmypid());
-    
-    switch ($sig) {
-      case SIGINT: $this->terminate= true; break;
-      case SIGHUP: $this->restart= true; break;
-    }
-  }
-
-  /**
    * Terminate child processes
    *
    * @param   array children
@@ -62,15 +48,16 @@ class PreforkingServer extends Server implements Traceable {
    */
   protected function _killChildren(&$children, $signal= SIGHUP) {
     foreach ($children as $pid => $i) {
-      $this->cat && $this->cat->infof('Server #%d: Terminating child #%d with pid %d', getmypid(), $i, $pid);
+      $this->cat && $this->cat->infof('Server #%d: Terminating child #%d, pid %d with %d', getmypid(), $i, $pid, $signal);
       posix_kill($pid, $signal);
+      pcntl_signal_dispatch();
 
       if (SIGHUP == $signal) continue;
 
       pcntl_waitpid($pid, $status, WUNTRACED);
       $this->cat && $this->cat->warnf('Server #%d: Child %d died with exitcode %d', getmypid(), $pid, $status);
     }
-    
+
     $this->restart= false;
   }
 
@@ -79,10 +66,16 @@ class PreforkingServer extends Server implements Traceable {
    *
    */
   public function handleChild() {
+    $this->terminate= false;
     
     // Install child signal handler.
-    pcntl_signal(SIGINT, array($this, 'handleSignal'));
-    pcntl_signal(SIGHUP, array($this, 'handleSignal'));
+    $handler= function($sig) {
+      $this->cat && $this->cat->debugf('Received signal %d in child %d', $sig, getmypid());
+      $this->terminate= true;
+      $this->socket->close();
+    };
+    pcntl_signal(SIGINT, $handler);
+    pcntl_signal(SIGHUP, $handler);
 
     // Handle initialization of protocol. This is called once for 
     // every new child created.
@@ -99,29 +92,34 @@ class PreforkingServer extends Server implements Traceable {
       // fails, break out of the loop and terminate the server - this really
       // should not happen!
       do {
-        $socketSelectInterrupted = false;
+        pcntl_signal_dispatch();
+        if ($this->terminate) return;
+
         if (false === socket_select($read, $null, $null, $timeout)) {
+          pcntl_signal_dispatch();
+          if ($this->terminate) return;
 
           // If socket_select has been interrupted by a signal, it will return FALSE,
           // but no actual error occurred - so check for "real" errors before throwing
           // an exception. If no error has occurred, skip over to the socket_select again.
-          if (0 !== socket_last_error($this->socket->_sock)) {
-            throw new \peer\SocketException('Call to select() failed');
-          } else {
-            $socketSelectInterrupted = true;
-            if ($this->terminate || $this->restart) return;
+          if (0 !== socket_last_error($this->socket->getHandle())) {
+            throw new \peer\SocketException('Call to select() failed - ');
           }
+
+          $accepted= false;
+        } else {
+          $this->socket->setBlocking(false);
+          $m= $this->socket->accept();
+          $this->socket->setBlocking(true);
+
+          // The call to accept() will return false if there is no client available.
+          // This is the case if select() detects activity but another child has already
+          // accepted this client. In this case, retry.
+          $accepted= $m instanceof \peer\Socket;
         }
+
       // if socket_select was interrupted by signal, retry socket_select
-      } while ($socketSelectInterrupted);
-
-      $m= $this->socket->accept();
-
-      // Sanity check
-      if (!($m instanceof \peer\Socket)) {
-        $this->cat && $this->cat->warn('Accepted socket type error ', \xp::typeOf($m));
-        return;
-      }
+      } while (!$accepted);
 
       // Handle accepted socket
       if ($this->protocol instanceof \peer\server\protocol\SocketAcceptHandler) {
@@ -139,6 +137,9 @@ class PreforkingServer extends Server implements Traceable {
       // If meanwhile the server is about to be shut
       // down, break loop and disconnect the client.
       do {
+        pcntl_signal_dispatch();
+        $this->cat && $this->cat->debugf('Child #%d handling data on %s...', getmypid(), $m->toString());
+
         try {
           $this->protocol->handleData($m);
         } catch (\io\IOException $e) {
@@ -146,6 +147,8 @@ class PreforkingServer extends Server implements Traceable {
           break;
         }
       } while ($m->isConnected() && !$m->eof() && !$this->terminate);
+
+      $this->cat && $this->cat->debugf('Child #%d closing...', getmypid());
 
       $m->close();
       $this->protocol->handleDisconnect($m);
@@ -162,6 +165,7 @@ class PreforkingServer extends Server implements Traceable {
   /**
    * Service
    *
+   * @return void
    */
   public function service() {
     if (!$this->socket->isConnected()) return false;
@@ -179,8 +183,6 @@ class PreforkingServer extends Server implements Traceable {
         $i++;
       } else {                // Child
         $this->handleChild();
-
-        // Exit out of child
         exit();
       }
       if (sizeof($children) < $this->count) continue;
@@ -189,8 +191,17 @@ class PreforkingServer extends Server implements Traceable {
       // process id of the process we are running in) will cleanly shut
       // down this server. If this server is run within a thread (which
       // is recommended), a $thread->stop() will accomplish this.
-      pcntl_signal(SIGINT, array($this, 'handleSignal'));
-      pcntl_signal(SIGHUP, array($this, 'handleSignal'));
+      $terminate= function($sig) {
+        $this->cat && $this->cat->debugf('Received terminate signal %d in server #%d', $sig, getmypid());
+        $this->terminate= true;
+      };
+      pcntl_signal(SIGINT, $terminate);
+
+      $restart= function($sig) {
+        $this->cat && $this->cat->debugf('Received restart signal %d in server #%d', $sig, getmypid());
+        $this->restart= true;
+      };
+      pcntl_signal(SIGHUP, $restart);
       
       // Wait until we are supposed to terminate. This condition variable
       // is set to TRUE by the signal handler. Sleep a microsecond to decrease
@@ -198,21 +209,20 @@ class PreforkingServer extends Server implements Traceable {
       // still be able to catch the shutdown signal in realtime.
       $this->cat && $this->cat->debug('Server #'.getmypid().': Starting main loop, children:', $children);
       while (!$this->terminate) { 
-        
+        pcntl_signal_dispatch();
+
         // If we get SIGHUP restart child
         // processes gracefully.
         if ($this->restart) {
           $this->_killChildren($children, SIGHUP);
         }
-        
-        // Minimize cpu usage.
-        usleep(1);
 
         // If, meanwhile, we've been interrupted, break out of both loops.
         if ($this->terminate) break 2;
         
         // If one or more of our children terminated, remove them
         // from the process list and fork new ones.
+        usleep(100000);
         while (($pid= pcntl_waitpid(-1, $status, WNOHANG)) > 0) {
           $this->cat && $this->cat->warnf('Server #%d: Child %d died with exitcode %d', getmypid(), $pid, $status);
           unset($children[$pid]);